+++ docs/source/release_notes.rst
@@ -2,7 +2,7 @@ Release notes
 =============
 
 
-0.1.6 (18.01.2023)
+0.1.6 (19.01.2023)
 -----
 - Fixed issue with duplicate subunits for non-expressed LRs when `return_all_lrs`` is True
 - `min_prop` when working with `return_all_lrs` is now filled with 0s
@@ -10,6 +10,8 @@ Release notes
 - Added `dotplot_by_sample` function to visualize ligand-receptor interactions across samples
 - Refractored of `dotplot` preprocessing
 - Changed "pvals" of geometric_mean method to "gmean_pvals" for consistency
+- `to_tensor_c2c` utility function to convert a long-format dataframe of ligand-receptor interactions by sample to Tensor-cell2cell tensor.
+- Added instances to `MethodMeta` class
 
 
 0.1.5 (11.01.2023)